Rewrite scp tests with testify

+++ lib/sshutils/scp/stat_linux.go
@@ -29,5 +29,5 @@ func atime(fi os.FileInfo) time.Time {
 }
 
 func timespecToTime(ts syscall.Timespec) time.Time {
-	return time.Unix(int64(ts.Sec), int64(ts.Nsec))
+	return time.Unix(ts.Sec, ts.Nsec)
 }
